rename: ImageCreateRequestDto 위치 변경 및 이름 변경
