+++ src/main/java/swyp/swyp6_team7/image/dto/ImageCreateDto.java
@@ -1,4 +1,4 @@
-package swyp.swyp6_team7.image.dto.request;
+package swyp.swyp6_team7.image.dto;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -7,7 +7,7 @@ import swyp.swyp6_team7.image.domain.Image;
 import java.time.LocalDateTime;
 
 @Getter
-public class ImageCreateRequestDto {
+public class ImageCreateDto {
 
     private String originalName; // 원본 파일 이름
     private String storageName; // 저장된 파일 이름 (고유 이름)
@@ -21,7 +21,7 @@ public class ImageCreateRequestDto {
     private LocalDateTime uploadDate;
 
     @Builder
-    public ImageCreateRequestDto(
+    public ImageCreateDto(
             String originalName, String storageName, long size, String format,
             String relatedType, int relatedNumber, int order, String key, String url, LocalDateTime uploadDate
     ) {
@@ -39,7 +39,7 @@ public class ImageCreateRequestDto {
 
     // 최소 필드만 받는 생성자 (프로필 이미지 기본 이미지용)
     @Builder
-    public ImageCreateRequestDto(String relatedType, int relatedNumber, int order, String key, String url) {
+    public ImageCreateDto(String relatedType, int relatedNumber, int order, String key, String url) {
         this.relatedType = relatedType;
         this.relatedNumber = relatedNumber;
         this.order = order;
@@ -50,7 +50,7 @@ public class ImageCreateRequestDto {
 
     //임시저장시 사용하는 생성자 (커뮤니티)
     @Builder
-    public ImageCreateRequestDto(
+    public ImageCreateDto(
             String originalName, String storageName, long size, String format,
             String relatedType, String key, String url
     ) {
